backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
psycopg2-binary>=2.9.9
python-multipart>=0.0.6
aiofiles>=23.2.0
alembic>=1.12.0

# LangChain for LLM integration
langchain>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.1.0

# PDF processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
pdfminer.six>=20221105

# DOCX processing
python-docx>=1.1.0

# OCR
paddleocr>=2.7.0
pytesseract>=0.3.10
Pillow>=10.0.0

# HTML parsing
beautifulsoup4>=4.12.0

# HTTP client for external APIs
httpx>=0.25.0

# Machine learning for layout detection
numpy>=1.24.0
scikit-learn>=1.3.0

# Additional utilities
requests>=2.31.0

# Pydantic EmailStr dependency
email-validator>=2.1.0.post1
